Renamed some chrome API files
