+++ src/browser/index.ts
@@ -14,14 +14,14 @@ export * from "./history";
 export * from "./i18n";
 export * from "./idle";
 export * from "./management";
-export * from "./notification";
+export * from "./notifications";
 export * from "./offscreen";
 export * from "./permissions";
 export * from "./runtime";
 export * from "./scripting";
 export * from "./sidebar";
 export * from "./tabCapture";
-export * from "./tab";
+export * from "./tabs";
 export * from "./userScripts";
 export * from "./webNavigation";
 export * from "./webRequest";
